Add support for yield*
